femora-backend/backend/requirements.txt
fastapi
uvicorn
pydantic
firebase-admin
pyjwt
python-dotenv
passlib[bcrypt]
